pyproject.toml
[tool]
[tool.poetry]
name = "centraal_client_flow"
version = "0.1.0"
homepage = "https://github.com/centraal-api/centraal_client_flow"
description = "Proyecto que facilita el flujo de información de cliente."
authors = ["German <equipo@centraal.studio>"]
readme = "README.md"
license =  "Apache-2.0"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English, Spanish',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
packages = [
    { include = "centraal_client_flow" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"

black  = { version = "^22.3.0", optional = true}
isort  = { version = "5.10.1", optional = true}
flake8  = { version = "4.0.1", optional = true}
flake8-docstrings = { version = "^1.6.0", optional = true }
pytest  = { version = "^8.0.1", optional = true}
pytest-cov  = { version = "^3.0.0", optional = true}
tox  = { version = "^3.24.5", optional = true}
mkdocs  = { version = "^1.2.3", optional = true}
mkdocs-include-markdown-plugin  = { version = "^3.2.3", optional = true}
mkdocs-material  = { version = "^8.1.11", optional = true}
mkdocstrings  = { version = "^0.18.0", optional = true}
mkdocs-material-extensions  = { version = "^1.0.3", optional = true}
twine  = { version = "^3.8.0", optional = true}
mkdocs-autorefs = {version = "^0.3.1", optional = true}
pre-commit = {version = "^2.17.0", optional = true}
toml = {version = "^0.10.2", optional = true}
livereload = {version = "^2.6.3", optional = true}
pyreadline = {version = "^2.1", optional = true}
mike = { version="^1.1.2", optional=true}
mypy = {version = "^1.5.1", optional = true}
setuptools = {version="^68.0", optional = true}
pkginfo = {version="^1.9", optional = true}
virtualenv = {version="^20.0", optional = true}
azure-cosmos  = {version="^4.7.0", optional = false}
pydantic  = {version="^2.8.0", optional = false}


[tool.poetry.extras]
test = [
    "pytest",
    "pytest-cov"
    ]

dev = ["tox", 
    "pre-commit", 
    "twine", 
    "toml", 
    "black",
    "isort",
    "flake8",
    "flake8-docstrings",
    "mypy"
    ]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs",
    "mike",
    "setuptools",
    "pkginfo",
    "virtualenv"
    ]



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | \.history
)/
'''
[tool.isort]
profile = "black"
